fix: read integer attribute value_bytes as uint32 LE, not float LE
getItemAttrValue was calling buf.readFloatLE(0) as the fallback when
attr.value is null. For integer attributes (e.g. killstreak tier 2025),
value_bytes stores the integer as uint32 LE — [0x02,0x00,0x00,0x00] = 2.
readFloatLE gives ~2.8e-45, causing weapon slot matching to silently fail
and the completeness check to reject a valid Mode A trade.

User-provided weapons sent via trade have the KS tier only in value_bytes
(not in the value field), while bot-owned weapons happen to have value=2
in the proto wire data — which is why the standalone worked but the
integrated bot did not.

Also adds:
- debug log of decoded fabricator slots before crafting
- slot-specific detail in the completeness check error message
- warning when backpack diff captures unexpected item count
- CLAUDE.md section on GC attribute data type gotchas

+++ src/lib/fabricatorSlots.ts
@@ -73,7 +73,7 @@ function getItemAttrValue(item: GCBackpackItem, attrDefIndex: number): number | 
     if (!attr) return null;
     if (attr.value !== null && attr.value !== undefined) return Number(attr.value);
     const buf = toBuffer(attr.value_bytes);
-    if (buf && buf.length >= 4) return buf.readFloatLE(0);
+    if (buf && buf.length >= 4) return buf.readUInt32LE(0);
     return null;
 }
